+++ sdk/keyvault/arm-keyvault/src/classic/operations/index.ts
@@ -12,13 +12,11 @@ export interface OperationsOperations {
   /** List the operations for the provider */
   list: (options?: OperationsListOptionalParams) => PagedAsyncIterableIterator<Operation>;
 }
-
 function _getOperations(context: KeyVaultManagementContext) {
   return {
     list: (options?: OperationsListOptionalParams) => list(context, options),
   };
 }
-
 export function _getOperationsOperations(context: KeyVaultManagementContext): OperationsOperations {
   return {
     ..._getOperations(context),