+++ sdk/keyvault/arm-keyvault/src/models/models.ts
@@ -1,15 +1,15 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { areAllPropsUndefined } from "../static-helpers/serialization/check-prop-undefined.js";
-import { uint8ArrayToString, stringToUint8Array } from "@azure/core-util";
-
-/**
+/*
  * This file contains only generated model types and their (de)serializers.
  * Disable the following rules for internal models with '_' prefix and deserializers which require 'any' for raw JSON input.
  */
 /* eslint-disable @typescript-eslint/naming-convention */
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+import { areAllPropsUndefined } from "../static-helpers/serialization/check-prop-undefined.js";
+import { uint8ArrayToString, stringToUint8Array } from "@azure/core-util";
+
 /** Result of the request to list Storage operations. It contains a list of operations and a URL link to get the next set of results. */
 export interface _OperationListResult {
   /** List of Storage operations supported by the Storage resource provider. */
@@ -306,6 +306,8 @@ export interface VaultProperties {
   readonly privateEndpointConnections?: PrivateEndpointConnectionItem[];
   /** Property to specify whether the vault will accept traffic from public internet. If set to 'disabled' all traffic except private endpoint traffic and that that originates from trusted services will be blocked. This will override the set firewall rules, meaning that even if the firewall rules are present we will not honor the rules. */
   publicNetworkAccess?: string;
+  /** Configuration for Token Binding for Entra tokens */
+  tokenBindingParameters?: TokenBindingParameters;
 }
 
 export function vaultPropertiesSerializer(item: VaultProperties): any {
@@ -329,6 +331,9 @@ export function vaultPropertiesSerializer(item: VaultProperties): any {
       : networkRuleSetSerializer(item["networkAcls"]),
     provisioningState: item["provisioningState"],
     publicNetworkAccess: item["publicNetworkAccess"],
+    tokenBindingParameters: !item["tokenBindingParameters"]
+      ? item["tokenBindingParameters"]
+      : tokenBindingParametersSerializer(item["tokenBindingParameters"]),
   };
 }
 
@@ -357,6 +362,9 @@ export function vaultPropertiesDeserializer(item: any): VaultProperties {
       ? item["privateEndpointConnections"]
       : privateEndpointConnectionItemArrayDeserializer(item["privateEndpointConnections"]),
     publicNetworkAccess: item["publicNetworkAccess"],
+    tokenBindingParameters: !item["tokenBindingParameters"]
+      ? item["tokenBindingParameters"]
+      : tokenBindingParametersDeserializer(item["tokenBindingParameters"]),
   };
 }
 
@@ -881,8 +889,8 @@ export interface PrivateEndpoint {
   readonly id?: string;
 }
 
-export function privateEndpointSerializer(item: PrivateEndpoint): any {
-  return item;
+export function privateEndpointSerializer(_item: PrivateEndpoint): any {
+  return {};
 }
 
 export function privateEndpointDeserializer(item: any): PrivateEndpoint {
@@ -990,11 +998,72 @@ export enum KnownPrivateEndpointConnectionProvisioningState {
  */
 export type PrivateEndpointConnectionProvisioningState = string;
 
+/** Configuration for Token Binding for Entra tokens */
+export interface TokenBindingParameters {
+  /** This specifies whether token binding is disabled, enabled or enforced. */
+  mode?: TokenBindingMode;
+  /** Must be one of the following values "NoValidation", "Unattested", "AttestedTrustedLaunch", "AttestedConfidential". Strength of the token binding increases with each value in that order. */
+  minimumTokenBindingStrength?: TokenBindingStrength;
+}
+
+export function tokenBindingParametersSerializer(item: TokenBindingParameters): any {
+  return { mode: item["mode"], minimumTokenBindingStrength: item["minimumTokenBindingStrength"] };
+}
+
+export function tokenBindingParametersDeserializer(item: any): TokenBindingParameters {
+  return {
+    mode: item["mode"],
+    minimumTokenBindingStrength: item["minimumTokenBindingStrength"],
+  };
+}
+
+/** This specifies whether token binding is disabled, enabled or enforced. */
+export enum KnownTokenBindingMode {
+  /** Token binding is enforced for the vault. Only bounded tokens will be accepted. Bearer tokens will be rejected. */
+  Enforced = "Enforced",
+  /** Token binding is not enforced for the vault. Bounded tokens will be rejected. */
+  NotEnforced = "NotEnforced",
+}
+
+/**
+ * This specifies whether token binding is disabled, enabled or enforced. \
+ * {@link KnownTokenBindingMode} can be used interchangeably with TokenBindingMode,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Enforced**: Token binding is enforced for the vault. Only bounded tokens will be accepted. Bearer tokens will be rejected. \
+ * **NotEnforced**: Token binding is not enforced for the vault. Bounded tokens will be rejected.
+ */
+export type TokenBindingMode = string;
+
+/** Must be one of the following values "NoValidation", "Unattested", "AttestedTrustedLaunch", "AttestedConfidential". Strength of the token binding increases with each value in that order. */
+export enum KnownTokenBindingStrength {
+  /** This is default when token binding is not enabled. */
+  NoValidation = "NoValidation",
+  /** No attestation proof is required for the bounded token. */
+  Unattested = "Unattested",
+  /** Bounded Entra token must originate from a trusted launch VM with attestation proof from the attestation authority like Microsoft Azure Attestation. */
+  AttestedTrustedLaunch = "AttestedTrustedLaunch",
+  /** Bounded Entra token must originate from a confidential VM with attestation proof from the attestation authority like Microsoft Azure Attestation. */
+  AttestedConfidential = "AttestedConfidential",
+}
+
+/**
+ * Must be one of the following values "NoValidation", "Unattested", "AttestedTrustedLaunch", "AttestedConfidential". Strength of the token binding increases with each value in that order. \
+ * {@link KnownTokenBindingStrength} can be used interchangeably with TokenBindingStrength,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **NoValidation**: This is default when token binding is not enabled. \
+ * **Unattested**: No attestation proof is required for the bounded token. \
+ * **AttestedTrustedLaunch**: Bounded Entra token must originate from a trusted launch VM with attestation proof from the attestation authority like Microsoft Azure Attestation. \
+ * **AttestedConfidential**: Bounded Entra token must originate from a confidential VM with attestation proof from the attestation authority like Microsoft Azure Attestation.
+ */
+export type TokenBindingStrength = string;
+
 /** The resource model definition for a Azure Resource Manager proxy resource. It will not have tags and a location */
 export interface ProxyResource extends Resource {}
 
-export function proxyResourceSerializer(item: ProxyResource): any {
-  return item;
+export function proxyResourceSerializer(_item: ProxyResource): any {
+  return {};
 }
 
 export function proxyResourceDeserializer(item: any): ProxyResource {
@@ -1020,8 +1089,8 @@ export interface Resource {
   readonly systemData?: SystemData;
 }
 
-export function resourceSerializer(item: Resource): any {
-  return item;
+export function resourceSerializer(_item: Resource): any {
+  return {};
 }
 
 export function resourceDeserializer(item: any): Resource {
@@ -1151,6 +1220,8 @@ export interface VaultPatchProperties {
   networkAcls?: NetworkRuleSet;
   /** Property to specify whether the vault will accept traffic from public internet. If set to 'disabled' all traffic except private endpoint traffic and that that originates from trusted services will be blocked. This will override the set firewall rules, meaning that even if the firewall rules are present we will not honor the rules. */
   publicNetworkAccess?: string;
+  /** Configuration for Token Binding for Entra tokens */
+  tokenBindingParameters?: TokenBindingParameters;
 }
 
 export function vaultPatchPropertiesSerializer(item: VaultPatchProperties): any {
@@ -1172,6 +1243,9 @@ export function vaultPatchPropertiesSerializer(item: VaultPatchProperties): any 
       ? item["networkAcls"]
       : networkRuleSetSerializer(item["networkAcls"]),
     publicNetworkAccess: item["publicNetworkAccess"],
+    tokenBindingParameters: !item["tokenBindingParameters"]
+      ? item["tokenBindingParameters"]
+      : tokenBindingParametersSerializer(item["tokenBindingParameters"]),
   };
 }
 
@@ -1936,8 +2010,8 @@ export interface MhsmPrivateEndpoint {
   readonly id?: string;
 }
 
-export function mhsmPrivateEndpointSerializer(item: MhsmPrivateEndpoint): any {
-  return item;
+export function mhsmPrivateEndpointSerializer(_item: MhsmPrivateEndpoint): any {
+  return {};
 }
 
 export function mhsmPrivateEndpointDeserializer(item: any): MhsmPrivateEndpoint {
@@ -2040,7 +2114,7 @@ export interface ManagedHsmSku {
   /** SKU Family of the managed HSM Pool */
   family: ManagedHsmSkuFamily;
   /** SKU of the managed HSM Pool */
-  name: ManagedHsmSkuName;
+  name: ManagedHsmSkuNameV2;
 }
 
 export function managedHsmSkuSerializer(item: ManagedHsmSku): any {
@@ -2071,9 +2145,33 @@ export enum KnownManagedHsmSkuFamily {
  * **C**
  */
 export type ManagedHsmSkuFamily = string;
+
 /** SKU of the managed HSM Pool */
-export type ManagedHsmSkuName =
-  "Standard_B1" | "Custom_B32" | "Custom_B6" | "Custom_C42" | "Custom_C10";
+export enum KnownManagedHsmSkuNameV2 {
+  /** Standard_B1 SKU */
+  StandardB1 = "Standard_B1",
+  /** Custom_B32 SKU */
+  CustomB32 = "Custom_B32",
+  /** Custom_B6 SKU */
+  CustomB6 = "Custom_B6",
+  /** Custom_C42 SKU */
+  CustomC42 = "Custom_C42",
+  /** Custom_C10 SKU */
+  CustomC10 = "Custom_C10",
+}
+
+/**
+ * SKU of the managed HSM Pool \
+ * {@link KnownManagedHsmSkuNameV2} can be used interchangeably with ManagedHsmSkuNameV2,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Standard_B1**: Standard_B1 SKU \
+ * **Custom_B32**: Custom_B32 SKU \
+ * **Custom_B6**: Custom_B6 SKU \
+ * **Custom_C42**: Custom_C42 SKU \
+ * **Custom_C10**: Custom_C10 SKU
+ */
+export type ManagedHsmSkuNameV2 = string;
 
 /** Managed service identity (system assigned and/or user assigned identities) */
 export interface ManagedServiceIdentity {
@@ -2139,8 +2237,8 @@ export interface UserAssignedIdentity {
   readonly clientId?: string;
 }
 
-export function userAssignedIdentitySerializer(item: UserAssignedIdentity): any {
-  return item;
+export function userAssignedIdentitySerializer(_item: UserAssignedIdentity): any {
+  return {};
 }
 
 export function userAssignedIdentityDeserializer(item: any): UserAssignedIdentity {
@@ -2679,6 +2777,8 @@ export enum KnownJsonWebKeyType {
   RSA = "RSA",
   /** RSA-HSM */
   RSAHSM = "RSA-HSM",
+  /** oct-HSM */
+  OctHSM = "oct-HSM",
 }
 
 /**
@@ -2689,7 +2789,8 @@ export enum KnownJsonWebKeyType {
  * **EC** \
  * **EC-HSM** \
  * **RSA** \
- * **RSA-HSM**
+ * **RSA-HSM** \
+ * **oct-HSM**
  */
 export type JsonWebKeyType = string;
 
@@ -3551,6 +3652,8 @@ export enum KnownVersions {
   V20250501 = "2025-05-01",
   /** The 2026-02-01 API version. */
   V20260201 = "2026-02-01",
+  /** The 2026-03-01-preview API version. */
+  V20260301Preview = "2026-03-01-preview",
 }
 
 export function _operationOperationPropertiesDeserializer(item: any) {